--- client/src/pages/roomReservation/hourBlock/HourBlock.module.css
@@ -1,17 +1,0 @@
-.container {
-  width: 100%;
-  border: 2px solid var(--border-100);
-  display: flex;
-  flex-direction: row;
-}
-
-.timeSlotWrapper {
-  position: relative;
-}
-
-.timeLabel {
-  position: absolute;
-  top: 0px;
-  left: 0;
-  font-size: 12px;
-}
