--- client/src/pages/roomReservation/timeSlot/TimeSlot.module.css
@@ -1,11 +1,0 @@
-.container {
-  width: 80%;
-  min-width: 30px;
-  border-right: 1px solid var(--text-200);
-  background-color: var(--bg-100);
-  flex-direction: row;
-  height: 90px;
-  justify-content: center;
-  align-items: center;
-}
-
